requirements.txt
defusedxml
pyyaml>=5.4